requirements.txt
Flask
Flask-MySQLdb
Flask-SQLAlchemy
PyMySQL
Flask-WTF
Flask-Bcrypt
